Make tests more indipendent from each other
