+++ src/c/tests/main_test.c
@@ -6,20 +6,10 @@ void start_tests()
 {
     start_session();
     
-    blank_memory();
     start_battle_tests();
-    
-    blank_memory();
     
     end_session();
     
     while(1)
         vblank();
-}
-
-void blank_memory()
-{
-    int blank_value = 0;
-    cpufastset(&blank_value, (void*)IWRAM, CPUFASTSET_FILL | (IWRAM_SIZE >> 2));
-    reg_ram_reset(NON_IWRAM_RESET);
 }
